run notebook conversion and git commands in kube job for ee
