+++ hopsworks-common/src/main/java/io/hops/hopsworks/common/git/CommandExecutor.java
@@ -1,0 +1,29 @@
+/*
+ * This file is part of Hopsworks
+ * Copyright (C) 2024, Hopsworks AB. All rights reserved
+ *
+ * Hopsworks is free software: you can redistribute it and/or modify it under the terms of
+ * the GNU Affero General Public License as published by the Free Software Foundation,
+ * either version 3 of the License, or (at your option) any later version.
+ *
+ * Hopsworks is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY;
+ * without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR
+ * PURPOSE.  See the GNU Affero General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License along with this program.
+ * If not, see <https://www.gnu.org/licenses/>.
+ */
+package io.hops.hopsworks.common.git;
+
+import com.logicalclocks.servicediscoverclient.exceptions.ServiceDiscoveryException;
+import io.hops.hopsworks.exceptions.GitOpException;
+import io.hops.hopsworks.persistence.entity.git.GitOpExecution;
+
+import java.io.IOException;
+
+public interface CommandExecutor {
+  void execute(GitOpExecution execution, BasicAuthSecrets authSecrets) throws GitOpException, IOException,
+    ServiceDiscoveryException;
+  void cancelGitExecution(GitOpExecution execution, String message);
+  void monitorCommands(GitOpExecution execution, String localMemberIp, boolean amIPrimary, long waitTime);
+}
