+++ hopsworks-common/src/main/java/io/hops/hopsworks/common/git/GitTimeoutCommandsMonitor.java
@@ -17,12 +17,9 @@ package io.hops.hopsworks.common.git;
 
 import io.hops.hopsworks.common.dao.git.GitOpExecutionFacade;
 import io.hops.hopsworks.common.dao.git.GitRepositoryFacade;
-import io.hops.hopsworks.common.git.util.GitCommandOperationUtil;
 import io.hops.hopsworks.common.util.PayaraClusterManager;
-import io.hops.hopsworks.common.util.Settings;
 import io.hops.hopsworks.persistence.entity.git.GitOpExecution;
 import io.hops.hopsworks.persistence.entity.git.GitRepository;
-import io.hops.hopsworks.persistence.entity.git.config.GitOpExecutionState;
 
 import javax.annotation.PostConstruct;
 import javax.annotation.PreDestroy;
@@ -35,6 +32,7 @@ import javax.ejb.Timeout;
 import javax.ejb.Timer;
 import javax.ejb.TimerConfig;
 import javax.ejb.TimerService;
+import javax.inject.Inject;
 import java.util.Collection;
 import java.util.Optional;
 import java.util.logging.Level;
@@ -46,7 +44,7 @@ import java.util.logging.Logger;
 public class GitTimeoutCommandsMonitor {
   private static final Logger LOGGER = Logger.getLogger(GitTimeoutCommandsMonitor.class.getName());
   //Time to wait to kill the container after it has reached timeout. 90 seconds
-  private static final Integer BONUS_TIME = 1000 * 90;
+  private static final long BONUS_TIME = 1000 * 90;
   //Time to wait after the repository is locked and the execution is created. 30s;
   private static final Integer WAIT_TIME_BEFORE_EXECUTION_OBJECT_CREATION = 3000;
 
@@ -55,11 +53,9 @@ public class GitTimeoutCommandsMonitor {
   @EJB
   private GitRepositoryFacade gitRepositoryFacade;
   @EJB
-  private GitCommandOperationUtil gitCommandOperationUtil;
-  @EJB
-  private Settings settings;
-  @EJB
   private PayaraClusterManager payaraClusterManager;
+  @Inject
+  private CommandExecutor commandExecutor;
 
   @Resource
   private TimerService timerService;
@@ -86,23 +82,15 @@ public class GitTimeoutCommandsMonitor {
   public void gitCommandMonitor(Timer timer) {
     //Should run on all
     LOGGER.log(Level.FINE, "Running GitTimeoutCommandsMonitor");
+    boolean amIPrimary = payaraClusterManager.amIThePrimary();
     Collection<GitRepository> repositories = gitRepositoryFacade.findAllWithOngoingOperations();
     for (GitRepository repository : repositories) {
       Optional<GitOpExecution> optional = gitOpExecutionFacade.findRunningInRepository(repository);
       if (optional.isPresent()) {
         GitOpExecution execution = optional.get();
         //we need all GitOpExecution to check for repository with a pid but no execution. So check host here
-        if (execution.getHostname().equals(localMemberIp)) {
-          long timeElapsed = System.currentTimeMillis() - execution.getExecutionStart();
-          if (timeElapsed > (settings.getGitJwtExpMs() + BONUS_TIME)) {
-            //kill this container
-            LOGGER.log(Level.INFO, "Killing git execution with Id + [{0}] with state {1}",
-              new Object[]{execution.getId(), execution.getState().toString()});
-            gitOpExecutionFacade.updateState(execution, GitOpExecutionState.TIMEDOUT, "Timeout");
-            gitCommandOperationUtil.shutdownCommandService(repository, execution);
-          }
-        }
-      } else if (payaraClusterManager.amIThePrimary()) {
+        commandExecutor.monitorCommands(execution, localMemberIp, amIPrimary, BONUS_TIME);
+      } else if (amIPrimary && repository.getCid() != null) {
         //A repository with a pid but no execution object
         try {
           long executionStart = Long.parseLong(repository.getCid());
