+++ hopsworks-common/src/main/java/io/hops/hopsworks/common/system/job/SystemJobStatus.java
@@ -1,0 +1,52 @@
+/*
+ * This file is part of Hopsworks
+ * Copyright (C) 2024, Hopsworks AB. All rights reserved
+ *
+ * Hopsworks is free software: you can redistribute it and/or modify it under the terms of
+ * the GNU Affero General Public License as published by the Free Software Foundation,
+ * either version 3 of the License, or (at your option) any later version.
+ *
+ * Hopsworks is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY;
+ * without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR
+ * PURPOSE.  See the GNU Affero General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License along with this program.
+ * If not, see <https://www.gnu.org/licenses/>.
+ */
+package io.hops.hopsworks.common.system.job;
+
+public class SystemJobStatus {
+  private int exitCode;
+  private String log;
+
+  public SystemJobStatus() {
+  }
+
+  public SystemJobStatus(int exitCode) {
+    this.exitCode = exitCode;
+  }
+
+  public int getExitCode() {
+    return exitCode;
+  }
+
+  public void setExitCode(int exitCode) {
+    this.exitCode = exitCode;
+  }
+
+  public String getLog() {
+    return log;
+  }
+
+  public void setLog(String log) {
+    this.log = log;
+  }
+
+  @Override
+  public String toString() {
+    return "SystemJobStatus{" +
+      "exitCode=" + exitCode +
+      ", log='" + log + '\'' +
+      '}';
+  }
+}
